requirements.txt
# Core dependencies
celery[redis]==5.5.3
flower==2.0.1
beautifulsoup4==4.12.2
lxml==4.9.3

# PII Detection and Masking
spacy==3.8.7
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354

# LLM and Embeddings
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.29
langchain-google-genai==2.1.9
langchain-ollama==0.3.6

# Vector Database
chromadb==1.0.16

# Database
boto3==1.40.5

# Web scraping for privacy policy
playwright==1.54.0

# Data validation and parsing
pydantic[email]==2.11.7
pydantic-settings==2.10.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity
PyYAML==6.0.1

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Evaluation and testing
deepeval==3.3.6
